auth_service/requirements.txt
flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
msal==1.31.0
google-cloud-secret-manager==2.20.2
